requirements.txt
fastapi
uvicorn
pdfplumber
pandas
requests
python-dotenv